+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Types of changes
 - [Changed] remove patchwork/utf8
 - [Fixed] remove the disabled attribute from buttons if reset filter
 - [Fixed] fix filter values
+- [Fixed] fix image sizes for list and detail view
 
 ## [2.7.9](https://github.com/pdir/maklermodul-bundle/tree/2.7.9) – 2021-09-15